--- src/ccpa/types/CCPAConsentState.ts
@@ -1,3 +1,0 @@
-export interface CCPAConsentState {
-	doNotSell: boolean;
-}
