--- src/tcfv2/types/ConsentList.ts
@@ -1,3 +1,0 @@
-export interface ConsentList {
-	[key: string]: boolean;
-}
