--- src/tcfv2/types/TCEventStatusCode.ts
@@ -1,4 +1,0 @@
-export type TCEventStatusCode =
-	| 'tcloaded'
-	| 'cmpuishown'
-	| 'useractioncomplete';
