--- src/tcfv2/types/TCFv2ConsentState.ts
@@ -1,8 +1,0 @@
-import { ConsentList } from './ConsentList';
-import { TCEventStatusCode } from './TCEventStatusCode';
-
-export interface TCFv2ConsentState {
-	consents: ConsentList;
-	eventStatus: TCEventStatusCode;
-	vendorConsents: ConsentList;
-}
