--- src/types/PubData.ts
@@ -1,3 +1,0 @@
-export interface PubData {
-	browserId?: string;
-}
