--- src/types/SourcepointImplementation.ts
@@ -1,8 +1,0 @@
-import { PubData } from './PubData';
-import { WillShowPrivacyMessage } from './WillShowPrivacyMessage';
-
-export interface SourcepointImplementation {
-	init: (pubData?: PubData) => void;
-	willShowPrivacyMessage: WillShowPrivacyMessage;
-	showPrivacyManager: () => void;
-}
