--- src/types/WillShowPrivacyMessage.ts
@@ -1,1 +1,0 @@
-export type WillShowPrivacyMessage = () => Promise<boolean>;
