+++ src/types/ccpa.ts
@@ -1,3 +1,7 @@
+export interface CCPAConsentState {
+	doNotSell: boolean;
+}
+
 export interface CCPAData {
 	version: number;
 	uspString: string;
