+++ src/types/tcfv2/CustomVendorConsents.ts
